Move request forwarding to forward.c
